Feat: Add "Back to Presentation" notice for print mode
Introduce a non-printable "Back to Presentation" notice when in print mode. This update creates a fixed position notice in the browser to guide users after printing, offering a button that removes the "print-pdf" mode from the URL and reloads the page. Additionally, styled the notice for both light and dark modes, and ensured it and the logo are not printed by using media queries.